Move route active tests to database package
